requirements.txt
openai
python-decouple
peewee
python-telegram-bot